+++ CHANGELOG.md
@@ -7,8 +7,10 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ### Fixed
 * [`no-unknown-property`]: add `dialog` attributes ([#3436][] @ljharb)
+* [`no-arrow-function-lifecycle`]: when converting from an arrow, remove the semi and wrapping parens ([#3337][] @ljharb)
 
 [#3436]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3436
+[#3337]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3337
 
 ## [7.31.8] - 2022.09.08